Cleanup tests, add new cases, fix timestamp issue on resume and complex type support in mocks
